+++ tests/helpers__intf_name.py
@@ -1,4 +1,4 @@
-"""unittest helpers"""
+"""Tests helpers"""
 
 # ============================ CISCO IOS =============================
 # show run | include interface