+++ tests/test__package.py
@@ -1,4 +1,4 @@
-"""unittests package"""
+"""Testss package"""
 import re
 from pathlib import Path